debt - inline `iframe` into `base/iframe`
